feat(release): only prompt user to create plans for groups with touched projects
